# openapi.json
{
  "components": {
    "schemas": {
...
      "FitRequest": {
        "properties": {
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "years": {
            "default": 5,
            "maximum": 25.0,
            "minimum": 0.0,
            "title": "Years",
            "type": "integer"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "FitRequest",
        "type": "object"
      },
      "FitResponse": {
        "properties": {
          "cis": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Cis",
            "type": "object"
          },
          "diagnostics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Diagnostics",
            "type": "object"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "n_obs": {
            "title": "N Obs",
            "type": "integer"
          },
          "params": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "model_name",
          "params",
          "cis",
          "diagnostics",
          "n_obs"
        ],
        "title": "FitResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimulateRequest": {
        "properties": {
          "fit": {
            "$ref": "#/components/schemas/FitResponse"
          },
          "horizon": {
            "default": 126,
            "maximum": 2520.0,
            "minimum": 1.0,
            "title": "Horizon",
            "type": "integer"
          },
          "n_paths": {
            "default": 10000,
            "maximum": 200000.0,
            "minimum": 1.0,
            "title": "N Paths",
            "type": "integer"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "x0": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X0"
          }
        },
        "required": [
          "fit"
        ],
        "title": "SimulateRequest",
        "type": "object"
      },
      "SimulateResponse": {
        "properties": {
          "bands": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Bands",
            "type": "object"
          },
          "horizon": {
            "title": "Horizon",
            "type": "integer"
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "sample_paths": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Sample Paths",
            "type": "array"
          }
        },
        "required": [
          "bands",
          "sample_paths",
          "horizon",
          "n_paths"
        ],
        "title": "SimulateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantMind API",
    "version": "0.4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/models/{name}/fit": {
      "post": {
        "operationId": "fit_api_models__name__fit_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fit"
      }
    },
    "/api/models/{name}/simulate": {
      "post": {
        "operationId": "simulate_api_models__name__simulate_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate"
      }
    },
...
  }
}